sh/pre_config/requirements.txt
Babel==2.6.0
dicttoxml==1.7.4
pbr==5.1.3
ply==3.11
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
six==1.12.0
yaql==1.1.3
